+++ src/textual_enhanced/dialogs/modal_input.py
@@ -5,6 +5,7 @@
 from textual import on
 from textual.app import ComposeResult
 from textual.screen import ModalScreen
+from textual.suggester import Suggester
 from textual.widgets import Input
 
 
@@ -32,6 +33,10 @@ class ModalInput(ModalScreen[str | None]):
         placeholder: str | None = None,
         initial: str = "",
         classes: str | None = None,
+        password: bool = False,
+        suggester: Suggester | None = None,
+        title: str | None = None,
+        sub_title: str | None = None,
     ) -> None:
         """Initialise the object.
 
@@ -39,16 +44,37 @@ class ModalInput(ModalScreen[str | None]):
             placeholder: The placeholder text to use.
             initial: The initial value for the input.
             classes: The CSS classes of the modal input.
+            password: Whether the input is a password input.
+            suggester: The suggester to use for the input.
+            title: The title of the modal input.
+            sub_title: The subtitle of the modal input.
         """
         super().__init__(classes=classes)
         self._placeholder = placeholder or ""
         """The placeholder to use for the input."""
         self._initial = initial
         """The initial value for the input."""
+        self._password = password
+        """Whether the input is a password input."""
+        self._suggester = suggester
+        """The suggester to use for the input."""
+        self._title = title
+        """The title of the modal input."""
+        self._subtitle = sub_title
+        """The subtitle of the modal input."""
 
     def compose(self) -> ComposeResult:
         """Compose the input dialog."""
-        yield Input(self._initial, placeholder=self._placeholder)
+        yield (
+            control := Input(
+                self._initial,
+                placeholder=self._placeholder,
+                suggester=self._suggester,
+                password=self._password,
+            )
+        )
+        control.border_title = self._title
+        control.border_subtitle = self._subtitle
 
     @on(Input.Submitted)
     def accept_input(self) -> None: